[App.js][m]: Move components to single page
