+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/mac/Documents/Datopian/gift-publisher/src/index.js":"1","/Users/mac/Documents/Datopian/gift-publisher/src/components/UploadComponent/upload.js":"2"},{"size":780,"mtime":1607505197421,"results":"3","hashOfConfig":"4"},{"size":9158,"mtime":1607505732679,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"8"},"y5b2zm",{"filePath":"9","messages":"10","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/mac/Documents/Datopian/gift-publisher/src/index.js",[],[],"/Users/mac/Documents/Datopian/gift-publisher/src/components/UploadComponent/upload.js",[]]
+[{"/Users/mac/Documents/Datopian/gift-publisher/src/index.js":"1","/Users/mac/Documents/Datopian/gift-publisher/src/App.js":"2"},{"size":722,"mtime":1607959625799,"results":"3","hashOfConfig":"4"},{"size":7250,"mtime":1607959646822,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"y5b2zm",{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/mac/Documents/Datopian/gift-publisher/src/index.js",[],"/Users/mac/Documents/Datopian/gift-publisher/src/App.js",[]]
